requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.5
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
asyncpg==0.29.0
attrs==25.3.0
certifi==2025.4.26
frozenlist==1.6.2
greenlet==3.2.3
idna==3.10
magic-filter==1.0.12
multidict==6.4.4
propcache==0.3.1
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.1.0
SQLAlchemy==2.0.30
typing_extensions==4.14.0
yarl==1.20.0
fastapi==0.111.0
pytest==8.3.3
pytest-asyncio==0.24.0
uvicorn==0.30.3
aiosqlite==0.20.0
PyYAML==6.0.2
passlib[bcrypt]==1.7.4